# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agora API
  version: 0.1.0
  summary: Ankara Üniversitesi öğrenci bilgi sistemi ve öğrenme ortamı API'si
  description: |
    Agora API'nin sözleşmesidir (contract-first): uç noktalar önce burada tasarlanır,
    backend bu sözleşmeye göre yazılır, web istemcisi de buradan üretilir.

    **Hata yanıtları** RFC 9457 `application/problem+json` biçimindedir. İstemciler
    `code` alanına göre karar verir. `detail` kullanıcıya gösterilebilir ama
    değişebilir.

    **Kimlik doğrulama:** korumalı uç noktalar `Authorization: Bearer <access_token>`
    ister. Access token 15 dakika geçerlidir. Süresi dolunca `POST /api/v1/auth/refresh`
    ile yenisi alınır.

    **Yetkilendirme:** yetkiler token'a gömülmez, her korumalı istekte sunucuda güncel
    olarak çözülür. Rol değişikliği ya da hesabın askıya alınması anında etkilidir.
    Belirli bir yetki gerektiren uçlarda yetki `x-agora-permission` alanında yazar.

    **Parola değiştirme zorunluluğu:** `must_change_password` açık bir kullanıcı sadece
    kendi hesabıyla ilgili uçlara (`/api/v1/me...`) erişebilir. Diğer bütün korumalı
    uçlar `403 PASSWORD_CHANGE_REQUIRED` döner. İstemci kullanıcıyı parola değiştirme
    ekranına yönlendirmelidir.

    **İki adımlı doğrulama (MFA):** hesabında MFA açık kullanıcının girişi iki adımlıdır:
    `POST /api/v1/auth/login` parola doğruysa token yerine `mfa_required: true` ve kısa
...
    görünür.

...
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/mfa/verify:
    post:
      tags: [Kimlik doğrulama]
      summary: Girişin ikinci adımı
      description: |
        Girişte alınan `mfa_token` ile doğrulayıcı uygulamadaki 6 haneli kodu (`code`) ya
        da bir kurtarma kodunu (`recovery_code`) gönderir. İkisinden tam olarak biri
        verilmelidir. Başarılıysa oturum açılır, yanıt girişle aynıdır.

        `mfa_token` 5 dakika geçerlidir ve en fazla 5 kod denemesi kabul eder. Yanlış kod
        hesabın başarısız deneme sayacını artırır (yanlış parola gibi). Aynı kod ve aynı
        kurtarma kodu ikinci kez kabul edilmez. Token, girişin başladığı istemci türüyle
        (`X-Agora-Client`) kullanılmalıdır. Hız sınırı girişle ortaktır.
      operationId: verifyMfa
      security: []
      parameters:
        - $ref: "#/components/parameters/ClientHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mfa_token]
              properties:
                mfa_token: { type: string }
                code: { type: string, pattern: "^[0-9 ]{6,7}$", examples: ["492039"] }
                recovery_code: { type: string, maxLength: 32, examples: ["k7m2p-x9qaz"] }
      responses:
        "200": { $ref: "#/components/responses/Tokens" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            `INVALID_MFA_CODE` (kod hatalı, aynı token'la tekrar denenebilir) ya da
            `INVALID_MFA_TOKEN` (token geçersiz, süresi dolmuş ya da deneme hakkı bitmiş:
            kullanıcı parolasını yeniden girmeli)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: "`ACCOUNT_DISABLED`: hesap bu arada askıya alınmış"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429":
          description: "`ACCOUNT_LOCKED` ya da `RATE_LIMITED` (bkz. giriş)"
          headers:
            Retry-After:
              schema: { type: integer, minimum: 0 }
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /api/v1/me/mfa:
    get:
      tags: [Hesap]
      summary: İki adımlı doğrulama durumu
      operationId: getMyMfa
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Durum
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFAStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/me/mfa/setup:
    post:
      tags: [Hesap]
      summary: İki adımlı doğrulama kurulumunu başlatır
      description: |
        Yeni bir TOTP sırrı üretir. İstemci `otpauth_uri`'yi QR kod olarak gösterir,
        okutulamazsa `secret` elle girilir. Kurulum `POST /api/v1/me/mfa/enable` ile
        tamamlanana kadar MFA kapalıdır. Tamamlanmamış kurulum yeniden başlatılabilir,
        eski sır geçersiz olur.
      operationId: startMfaSetup
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sır üretildi
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFASetup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`MFA_ALREADY_ENABLED`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/me/mfa/enable:
    post:
      tags: [Hesap]
      summary: İki adımlı doğrulamayı açar
      description: |
        Mevcut parola ve doğrulayıcı uygulamadaki kod istenir. Yanlış parola başarısız
        deneme sayacını artırır. Başarılıysa 10 kurtarma kodu döner: kodlar sadece bu
        yanıtta görünür.

        Kullanıcının diğer bütün oturumları kapatılır. İsteği yapan oturum MFA'lı sayılır:
        istemci `POST /api/v1/auth/refresh` ile yeni access token almalıdır.
      operationId: enableMfa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, code]
              properties:
                current_password: { type: string, format: password }
                code: { type: string, examples: ["492039"] }
      responses:
        "200":
          description: MFA açıldı
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecoveryCodes" }
        "400":
          description: "`VALIDATION_FAILED`, `INVALID_CURRENT_PASSWORD` ya da `INVALID_MFA_CODE`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`MFA_ALREADY_ENABLED` ya da `MFA_SETUP_REQUIRED` (önce kurulum başlatılmalı)"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: "`ACCOUNT_LOCKED`"
          headers:
            Retry-After:
              schema: { type: integer, minimum: 0 }
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/me/mfa/disable:
    post:
      tags: [Hesap]
      summary: İki adımlı doğrulamayı kapatır
      description: |
        Mevcut parola ile doğrulayıcı uygulamadaki kod ya da bir kurtarma kodu istenir.
        Yanlış parola ve yanlış kod başarısız deneme sayacını artırır. Kullanıcının diğer
        oturumları kapatılır, isteği yapan oturum artık MFA'lı sayılmaz.
      operationId: disableMfa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password]
              properties:
                current_password: { type: string, format: password }
                code: { type: string }
                recovery_code: { type: string, maxLength: 32 }
      responses:
        "204":
          description: MFA kapandı
        "400":
          description: "`VALIDATION_FAILED`, `INVALID_CURRENT_PASSWORD` ya da `INVALID_MFA_CODE`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`MFA_NOT_ENABLED`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: "`ACCOUNT_LOCKED`"
          headers:
            Retry-After:
              schema: { type: integer, minimum: 0 }
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/me/mfa/recovery-codes:
    post:
      tags: [Hesap]
      summary: Kurtarma kodlarını yeniler
      description: |
        Doğrulayıcı uygulamadaki kod istenir (kurtarma kodu kabul edilmez). Eski kodların
        hepsi geçersiz olur, yeni 10 kod sadece bu yanıtta görünür.
      operationId: regenerateRecoveryCodes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: Yeni kodlar
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecoveryCodes" }
        "400":
          description: "`VALIDATION_FAILED` ya da `INVALID_MFA_CODE`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`MFA_NOT_ENABLED`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: "`ACCOUNT_LOCKED`"
          headers:
            Retry-After:
              schema: { type: integer, minimum: 0 }
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT (EdDSA)
  parameters:
    ClientHeader:
      name: X-Agora-Client
      in: header
      required: true
      description: Refresh token'ın çerezde (web) mi gövdede (mobile) mi taşınacağını belirler.
      schema:
        type: string
        enum: [web, mobile]
...
  responses:
    Tokens:
      description: Giriş veya yenileme başarılı
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
        Set-Cookie:
          description: |
            Sadece web istemcisinde:
            `agora_refresh=<token>; Path=/api/v1/auth; Max-Age=<saniye>; HttpOnly; Secure; SameSite=Strict`
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Tokens" }
    BadRequest:
      description: |
        `INVALID_CLIENT` (X-Agora-Client eksik ya da geçersiz), `INVALID_BODY`
        (JSON bozuk, bilinmeyen alan, yanlış tür) ya da `VALIDATION_FAILED` (alan hataları
        `errors` dizisinde)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        `AUTHENTICATION_REQUIRED` (Bearer token yok), `INVALID_TOKEN` (geçersiz veya
        süresi dolmuş) ya da `SESSION_REVOKED` (oturum çıkışla ya da güvenlik nedeniyle
        sonlandırılmış). İstemci bir kez refresh deneyip isteği tekrarlamalıdır. Refresh
        de başarısız olursa giriş ekranına yönlenmelidir.
      headers:
        WWW-Authenticate:
          schema: { type: string, examples: ['Bearer realm="agora", error="invalid_token"'] }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: |
        `FORBIDDEN` (gerekli yetki yok), `ACCOUNT_DISABLED` (hesap askıya alınmış veya
        kapatılmış), `PASSWORD_CHANGE_REQUIRED` (önce parola değiştirilmeli) ya da
        `MFA_REQUIRED` (yetki var ama bu oturum iki adımlı doğrulamayla açılmamış)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PayloadTooLarge:
      description: "`BODY_TOO_LARGE`: gövde 1 MiB'ı aşıyor"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    UnsupportedMediaType:
      description: "`UNSUPPORTED_MEDIA_TYPE`: gövde application/json değil"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, examples: ["about:blank"] }
        title: { type: string, examples: ["Unauthorized"] }
        status: { type: integer, examples: [401] }
        detail: { type: string }
        instance: { type: string, examples: ["/api/v1/auth/login"] }
        code: { type: string, examples: ["INVALID_CREDENTIALS"] }
        request_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string, description: "Kullanıcıya gösterilebilir (Türkçe)" }
              code: { type: string, description: "Makine okunur kod (varsa), ör. TOO_SHORT" }
...
    Tokens:
      type: object
      required: [access_token, token_type, expires_in, session_id, must_change_password]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in:
          type: integer
          description: Access token'ın kalan ömrü (saniye)
          examples: [900]
        refresh_token:
          type: string
          description: Sadece mobil istemcide
        refresh_expires_in:
          type: integer
          description: Sadece mobil istemcide. Refresh token'ın kalan ömrü (saniye)
        session_id: { type: string, format: uuid }
        must_change_password:
          type: boolean
          description: true ise istemci kullanıcıyı parola değiştirme ekranına yönlendirmeli
...
    MFAStatus:
      type: object
      required: [enabled, enabled_at, recovery_codes_remaining]
      properties:
        enabled: { type: boolean }
        enabled_at: { type: [string, "null"], format: date-time }
        recovery_codes_remaining: { type: integer, minimum: 0, maximum: 10 }
    MFASetup:
      type: object
      required: [secret, otpauth_uri]
      properties:
        secret:
          type: string
          description: Base32 (dolgusuz). QR okutulamazsa elle girilir
          examples: ["JBSWY3DPEHPK3PXPJBSWY3DPEHPK3PXP"]
        otpauth_uri:
          type: string
          description: QR kod olarak gösterilir (SHA1, 6 hane, 30 sn)
          examples: ["otpauth://totp/Agora:22290001?algorithm=SHA1&digits=6&issuer=Agora&period=30&secret=JBSW..."]
    RecoveryCodes:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          minItems: 10
          maxItems: 10
          items: { type: string, examples: ["k7m2p-x9qaz"] }
...